server-driven/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
anyascii==0.3.2
arrow==1.3.0
astroid==3.2.2
asttokens==2.4.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
audioread==3.0.1
autopep8==2.0.4
Babel==2.15.0
backports.tarfile==1.2.0
bangla==0.0.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
black==24.4.2
bleach==6.1.0
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2022.6.15
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.1.0
click==8.1.3
cloudpathlib==0.16.0
cloudpickle==3.0.0
comm==0.2.2
confection==0.1.4
contourpy==1.2.1
cookiecutter==2.6.0
coqpit==0.0.17
coqui-tts==0.24.1
coqui-tts-trainer==0.1.1
cryptography==42.0.7
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.8
docopt==0.6.2
docstring-to-markdown==0.15
docutils==0.21.2
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.14.0
flake8==7.0.0
Flask==2.2.2
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.5.0
g2pkk==0.1.2
grpcio==1.64.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gymnasium==0.29.1
hangul-romanize==0.1.0
huggingface-hub==0.23.2
idna==3.3
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflect==7.2.1
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.18.1
isort==5.13.2
itsdangerous==2.1.2
jamo==0.4.1
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
jellyfish==1.0.4
jieba==0.42.1
Jinja2==3.1.2
joblib==1.4.2
jsonlines==1.2.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keyring==25.2.1
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
marisa-trie==1.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
numpydoc==1.7.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
packaging==24.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
playsound==1.3.0
pluggy==1.5.0
pooch==1.8.1
preshed==3.0.9
prompt_toolkit==3.0.45
protobuf==5.27.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pycodestyle==2.9.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.18.0
pylint==3.2.2
pylint-venv==3.0.3
pyls-spyder==0.4.0
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==0.20.0
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
python-slugify==8.0.4
pytoolconfig==1.3.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.0.3
QDarkStyle==3.2.3
qstylizer==0.2.3
QtAwesome==1.3.1
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.5.15
requests==2.28.1
rich==13.7.1
rope==1.13.0
rpds-py==0.18.1
Rtree==1.2.0
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.11.4
SecretStorage==3.3.3
six==1.16.0
smart-open==6.4.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
spyder==5.5.4
spyder-kernels==2.5.1
srsly==2.4.8
stable_baselines3==2.3.2
stack-data==0.6.3
stripe==4.1.0
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
text-unidecode==1.3
textdistance==4.6.2
thinc==8.2.3
threadpoolctl==3.5.0
three-merge==0.1.1
tinycss2==1.3.0
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.5
torch==2.3.0
torchaudio==2.3.0
tornado==6.4
tqdm==4.66.4
trainer==0.0.36
traitlets==5.14.3
transformers==4.40.2
triton==2.3.0
TTS==0.22.0
typeguard==4.3.0
typer==0.9.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.0
tzlocal==5.2
ujson==5.10.0
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==1.26.11
wasabi==1.1.2
watchdog==4.0.1
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
Werkzeug==2.2.2
whatthepatch==1.0.5
wurlitzer==3.1.0
yapf==0.32.0
yarl==1.9.4
zipp==3.19.0
